added edit and settings features

+++ src/services/apiCabins.js
@@ -1,5 +1,6 @@
 import supabase, { supabaseUrl } from "./supabase";
 
+// Function to Get all the cabins table
 export async function getCabins() {
   const { data, error } = await supabase.from("cabins").select("*");
 
@@ -11,26 +12,22 @@ export async function getCabins() {
   return data;
 }
 
+// Function to delete the cabins row with that ID
 export async function deleteCabin(id) {
   const { error } = await supabase.from("cabins").delete().eq("id", id);
 
   if (error) {
-    console.log("Cabins could not be loaded");
+    console.log("Cabins could not be deleted");
     throw new Error("Cabins could not be deleted");
   }
 }
 
-export async function createCabin(newCabin) {
-  // This one works and is a different image in the table
+// Function to create or edit a cabin row
+export async function createEditCabin(newCabin, id) {
+  // Example image URL
   // https://kfozxafslhzhdihdnswx.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
 
-  // This is the new url gotten from supabase table row
-  // https://kfozxafslhzhdihdnswx.supabase.co/storage/v1/object/public/cabin-images/0.2344930367738649-cabin-008.jpg
-
-  // This one is gotten direct from the buckets and i used it directly in a table in supabase and it works
-  // https://kfozxafslhzhdihdnswx.supabase.co/storage/v1/object/public/cabin-images/imageName
-
-  // the last 2 images are the same
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
   // creating a random number for the image
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
@@ -39,13 +36,24 @@ export async function createCabin(newCabin) {
   );
 
   // Configuring the path to be saved in the bucket
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // 1. Create cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  // 1. Select the table
+  let query = supabase.from("cabins");
+
+  // A) Create if no valid Id
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) Edit
+  if (id)
+    query = query
+      .update({ ...newCabin, image: imagePath })
+      .eq("id", id)
+      .select();
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.log("Cabins could not be loaded");
@@ -53,10 +61,12 @@ export async function createCabin(newCabin) {
   }
 
   // 2. Upload image
+  if (hasImagePath) return data;
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
+  // If image wasn't successful delete the row
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.log(storageError);
